Use refit to refactor http requests on frontend
