+++ LibraryWebApp/Pages/Index.cshtml.cs
@@ -1,21 +1,21 @@
+using LibraryWebApp.Abstraction;
 using LibraryWebApp.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Refit;
 
 namespace LibraryWebApp.Pages
 {
     public class IndexModel : PageModel
     {
         private readonly ILogger<IndexModel> _logger;
-        private readonly HttpClient _httpClient;
+        private readonly IBookClient _bookClient;
         public List<Book> Books { get; set; } = new List<Book>();
         public Book? SelectedBook { get; set; }
 
-        public IndexModel(ILogger<IndexModel> logger, IHttpClientFactory httpClientFactory, IConfiguration configuration)
+        public IndexModel(ILogger<IndexModel> logger, IBookClient bookClient, IConfiguration configuration)
         {
             _logger = logger;
-            _httpClient = httpClientFactory.CreateClient("library_api");
+            _bookClient = bookClient;
         }
 
         public void OnGet()
@@ -25,14 +25,14 @@ namespace LibraryWebApp.Pages
         public async Task<IActionResult> OnPostBooks(string? bookTitleKey)
         {
             SelectedBook = null;
-            Books = await _httpClient.GetFromJsonAsync<List<Book>>($"api/books?key={bookTitleKey?.Trim() ?? string.Empty}") ?? new List<Book>();
+            Books = await _bookClient.Get(bookTitleKey?.Trim() ?? string.Empty);
             return Page();
         }
 
         public async Task<IActionResult> OnPostBooksByAuthor(string? authorNameKey)
         {
             SelectedBook = null;
-            Books = await _httpClient.GetFromJsonAsync<List<Book>>($"api/books?key=") ?? new List<Book>();
+            Books = await _bookClient.Get(string.Empty);
             authorNameKey = authorNameKey?.Trim() ?? string.Empty;
             Books = Books.Where(b => b.author.first_name.Contains(authorNameKey, StringComparison.InvariantCultureIgnoreCase) || b.author.last_name.Contains(authorNameKey, StringComparison.InvariantCultureIgnoreCase)).ToList();
             return Page();
@@ -41,7 +41,7 @@ namespace LibraryWebApp.Pages
         public async Task<IActionResult> OnPostBooksByPublisher(string? publisherNameKey)
         {
             SelectedBook = null;
-            Books = await _httpClient.GetFromJsonAsync<List<Book>>($"api/books?key=") ?? new List<Book>();
+            Books = await _bookClient.Get(string.Empty);
             publisherNameKey = publisherNameKey?.Trim() ?? string.Empty;
             Books = Books.Where(b => b.publisher.Contains(publisherNameKey, StringComparison.InvariantCultureIgnoreCase)).ToList();
             return Page();
